refactor: 폴더 성격에 맞게 파일 이동
