+++ src/pages/ReportEdit/Page.tsx
@@ -26,8 +26,8 @@ import { readReportDetail } from "@/apis/manager";
 import { ImageUploadApi as ImageUploadToServer } from "@/apis/rank";
 import { modifyReport, postReport } from "@/apis/report";
 import { getMyTeamUsers } from "@/apis/users";
-import Heic2Jpg from "@/utils/Image/Heic2Jpg";
-import compressedFile from "@/utils/Image/compressFile";
+import Heic2Jpg from "@/utils/Heic2Jpg";
+import compressedImageFile from "@/utils/compressImageFile";
 import { StudyCertificationDialog } from "@/pages/ReportAdd/components/StudyCertificationDialog";
 import {
   Form,
@@ -44,7 +44,7 @@ import { useCallback, useEffect, useRef } from "react";
 import { useQueries, useQuery } from "react-query";
 import { useNavigate, useParams } from "react-router-dom";
 import { toast } from "sonner";
-import { addImagePrefix } from "@/utils/Image/imagePrefix";
+import { addImagePrefix } from "@/utils/imagePrefix";
 import { WaveLoading } from "@/components/WaveLoading";
 import { NoData } from "@/components/NoData";
 
@@ -178,7 +178,7 @@ export default function ReportEditPage() {
 
     const convertedFile = await Heic2Jpg(file[0]);
 
-    const lowCapacityFile = await compressedFile(convertedFile as File);
+    const lowCapacityFile = await compressedImageFile(convertedFile as File);
 
     const reader = new FileReader();
     reader.onloadend = () => {
